move module `fauxhdfbuilder.py` from `bapsflib.lapd._hdf.tests` to `bapsflib._hdf.maps.tests`
