+++ bapsflib/_hdf/maps/controls/tests/common.py
@@ -14,7 +14,7 @@ import os
 import re
 import unittest as ut
 
-from bapsflib.lapd._hdf.tests import FauxHDFBuilder
+from bapsflib._hdf.maps import FauxHDFBuilder
 
 from .. import ConType
 from ..templates import (HDFMapControlTemplate,
