+++ bapsflib/_hdf/maps/digitizers/tests/test_map_digis.py
@@ -15,7 +15,7 @@ import h5py
 import numpy as np
 import unittest as ut
 
-from bapsflib.lapd._hdf.tests import FauxHDFBuilder
+from bapsflib._hdf.maps import FauxHDFBuilder
 
 from ..map_digis import HDFMapDigitizers
 from ..templates import HDFMapDigiTemplate
