+++ bapsflib/_hdf/maps/msi/tests/common.py
@@ -13,7 +13,7 @@ import numpy as np
 import os
 import unittest as ut
 
-from bapsflib.lapd._hdf.tests import FauxHDFBuilder
+from bapsflib._hdf.maps import FauxHDFBuilder
 
 from ..templates import HDFMapMSITemplate
