+++ bapsflib/lapd/_hdf/tests/__init__.py
@@ -8,4 +8,3 @@
 # License: Standard 3-clause BSD; see "LICENSES/LICENSE.txt" for full
 #   license terms and contributor agreement.
 #
-from .fauxhdfbuilder import FauxHDFBuilder
